shift swipefeature to utils
